Group device kernels by encoding-specificity
Align the module layout with the BlockEncoding architecture. New
common_kernels module holds the encoding-independent device kernels
(reflect_about_zero, controlled_reflect_about_zero, signal_phase,
controlled_signal_phase — geometry-only, valid for any zero-flagged
encoding). The LCU composite kernels move next to the rest of the LCU
circuits in pauli_lcu (walk steps, prepare reflections, and the
phase-sequence escape hatches), which also lets PauliLCU's factory
methods drop their deferred qubitization imports. qubitization and
qsvt now contain only the encoding-generic layer: Walk + observables,
PhaseSequence + QSVT. The unused FORWARD/ADJOINT constants in
qubitization are removed (qsvt owns them).

Package-level exports are unchanged; only submodule homes moved.
