+++ python/cudaq_algorithms/__init__.py
@@ -31,17 +31,18 @@ except ImportError:
     __version__ = "CUDA-Q Algorithms (pure Python; native extension not built)"
 
 # Pure-Python quantum primitives (no compiled-extension dependency).
-from . import block_encoding, pauli_lcu, qsvt, qubitization, sim_utils
+from . import (block_encoding, common_kernels, pauli_lcu, qsvt, qubitization,
+               sim_utils)
 from .block_encoding import BlockEncoding
-from .pauli_lcu import (PauliLCU, apply, controlled_select, prepare,
-                        reflect_about_zero, select, state_from, unprepare,
-                        walk)
+from .common_kernels import (controlled_reflect_about_zero,
+                             controlled_signal_phase, reflect_about_zero,
+                             signal_phase)
+from .pauli_lcu import (PauliLCU, adjoint_walk, apply,
+                        apply_controlled_phase_sequence, apply_phase_sequence,
+                        controlled_adjoint_walk,
+                        controlled_reflect_about_prepare, controlled_select,
+                        controlled_walk, prepare, reflect_about_prepare,
+                        select, state_from, unprepare, walk)
 from .qsvt import (ADJOINT, FORWARD, PhaseSequence, QSVT,
-                   apply_controlled_phase_sequence, apply_phase_sequence,
-                   controlled_signal_phase, recover_real_time_evolution,
-                   signal_phase)
-from .qubitization import (Walk, adjoint_walk, controlled_adjoint_walk,
-                           controlled_reflect_about_prepare,
-                           controlled_reflect_about_zero, controlled_walk,
-                           reflect_about_prepare, reflection_observable,
-                           select_observable)
+                   recover_real_time_evolution)
+from .qubitization import Walk, reflection_observable, select_observable
